# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Echobot",
    "description": "AI驱动的数字人动画平台 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/user/profile": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取用户资料",
        "description": "获取当前登录用户的个人资料\n\n返回用户的基本信息、订阅状态等。\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"6948bda91fd873cf81f5addd\",\n        \"email\": \"user@example.com\",\n        \"nickname\": \"小明\",\n        \"avatar_url\": \"https://example.com/avatar.jpg\",\n        \"role\": \"user\",\n        \"subscription\": {\"plan\": \"free\", \"expires_at\": null},\n        \"is_active\": true,\n        \"created_at\": \"2025-01-01T12:00:00\"\n    }\n}\n```",
        "operationId": "get_profile_api_v1_user_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "更新用户资料",
        "description": "更新当前用户的个人资料\n\n可更新的字段：\n- **nickname**: 用户昵称\n\n**请求示例:**\n```json\n{\n    \"nickname\": \"新昵称\"\n}\n```",
        "operationId": "update_profile_api_v1_user_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/user/avatar": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "上传头像",
        "description": "上传用户头像\n\n**支持的图片格式:** JPG, PNG, GIF, WebP\n\n**文件大小限制:** 最大 5MB\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"6948bda91fd873cf81f5addd\",\n        \"avatar_url\": \"/uploads/avatars/user123/avatar_20250101.jpg\",\n        ...\n    }\n}\n```",
        "operationId": "upload_avatar_api_v1_user_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_user_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/user/change-password": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "修改密码",
        "description": "修改当前用户的登录密码\n\n需要提供当前密码进行验证，新密码必须满足以下要求：\n- 长度至少8位\n- 包含至少一个大写字母\n- 包含至少一个小写字母\n- 包含至少一个数字\n\n**请求示例:**\n```json\n{\n    \"old_password\": \"OldPassword123\",\n    \"new_password\": \"NewPassword456\"\n}\n```\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"密码修改成功\",\n    \"data\": null\n}\n```",
        "operationId": "change_password_api_v1_user_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/stories": {
      "get": {
        "tags": [
          "故事"
        ],
        "summary": "获取故事列表",
        "description": "获取已发布的故事列表\n\n支持分页、分类筛选和关键词搜索。\n普通用户只能查看已发布状态的故事。\n\n**请求参数:**\n- **page**: 页码，从1开始\n- **page_size**: 每页数量，1-100\n- **category_id**: 分类ID（可选），筛选特定分类下的故事\n- **search**: 搜索关键词（可选），匹配故事标题\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"items\": [\n            {\n                \"id\": \"694534f1c32ffbd6151c18a2\",\n                \"title\": \"三只小猪\",\n                \"title_en\": \"The Three Little Pigs\",\n                \"category_id\": \"694534f1c32ffbd6151c18a1\",\n                \"thumbnail_url\": \"/uploads/thumbnails/story123.jpg\",\n                \"duration\": 180,\n                \"is_published\": true,\n                \"view_count\": 1234,\n                \"created_at\": \"2025-01-01T12:00:00\"\n            }\n        ],\n        \"total\": 50,\n        \"page\": 1,\n        \"page_size\": 20\n    }\n}\n```",
        "operationId": "list_stories_api_v1_stories_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从1开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从1开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量，最大100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量，最大100"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类ID，用于筛选特定分类",
              "title": "Category Id"
            },
            "description": "分类ID，用于筛选特定分类"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词，匹配标题",
              "title": "Search"
            },
            "description": "搜索关键词，匹配标题"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/random": {
      "get": {
        "tags": [
          "故事"
        ],
        "summary": "获取随机故事",
        "description": "获取随机故事列表\n\n随机返回指定数量的已发布故事，用于首页轮播图和推荐展示。\n\n**请求参数:**\n- **limit**: 返回数量，1-20，默认10\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": [\n        {\n            \"id\": \"694534f1c32ffbd6151c18a2\",\n            \"title\": \"三只小猪\",\n            \"thumbnail_url\": \"/uploads/thumbnails/story123.jpg\",\n            \"duration\": 180\n        }\n    ]\n}\n```",
        "operationId": "get_random_stories_api_v1_stories_random_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "返回数量，最大20",
              "default": 10,
              "title": "Limit"
            },
            "description": "返回数量，最大20"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/{story_id}": {
      "get": {
        "tags": [
          "故事"
        ],
        "summary": "获取故事详情",
        "description": "获取单个故事的完整详情\n\n返回故事的所有信息，包括视频URL、字幕数据等。\n\n**路径参数:**\n- **story_id**: 故事唯一标识ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"694534f1c32ffbd6151c18a2\",\n        \"title\": \"三只小猪\",\n        \"title_en\": \"The Three Little Pigs\",\n        \"category_id\": \"694534f1c32ffbd6151c18a1\",\n        \"category_name\": \"童话故事\",\n        \"description\": \"三只小猪建房子的经典童话故事\",\n        \"thumbnail_url\": \"/uploads/thumbnails/story123.jpg\",\n        \"video_url\": \"/uploads/videos/story123.mp4\",\n        \"audio_url\": \"/uploads/audio/story123.mp3\",\n        \"duration\": 180,\n        \"is_published\": true,\n        \"view_count\": 1234,\n        \"subtitles\": [\n            {\"start\": 0.0, \"end\": 3.5, \"text\": \"从前有三只小猪...\"}\n        ],\n        \"created_at\": \"2025-01-01T12:00:00\",\n        \"updated_at\": \"2025-01-01T12:00:00\"\n    }\n}\n```",
        "operationId": "get_story_api_v1_stories__story_id__get",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/{story_id}/view": {
      "post": {
        "tags": [
          "故事"
        ],
        "summary": "记录播放次数",
        "description": "记录故事播放次数\n\n每次用户观看故事时调用此接口，用于统计播放量。\n\n**路径参数:**\n- **story_id**: 故事唯一标识ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"播放记录成功\",\n    \"data\": null\n}\n```",
        "operationId": "record_view_api_v1_stories__story_id__view_post",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_avatar_api_v1_user_avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "头像图片文件，支持 JPG/PNG/GIF/WebP 格式"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_api_v1_user_avatar_post"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "old_password": {
            "type": "string",
            "minLength": 1,
            "title": "Old Password",
            "description": "当前密码",
            "example": "OldPassword123"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "description": "新密码，至少8位，必须包含大小写字母和数字",
            "example": "NewPassword456"
          }
        },
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "description": "修改密码请求\n\n用户修改自己的登录密码。\n需要提供旧密码进行验证，新密码需要满足强度要求。",
        "example": {
          "new_password": "NewPassword456",
          "old_password": "OldPassword123"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname",
            "description": "新昵称，2-50个字符，支持中文、字母、数字和下划线",
            "example": "新昵称"
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest",
        "description": "更新用户资料请求\n\n用于修改用户的个人信息，如昵称等。\n所有字段都是可选的，只更新提供的字段。",
        "example": {
          "nickname": "新昵称"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}